Fix horribly wrong defintions of ULONG32 and ULONG, as well as elf support which depended on these wrong definitions.
svn path=/trunk/; revision=12713

+++ reactos/lib/rtl/mem.c
@@ -1,5 +1,5 @@
 
-/* $Id: mem.c,v 1.3 2004/11/25 19:25:06 ekohl Exp $
+/* $Id$
  * 
  * COPYRIGHT:       See COPYING in the top level directory
  * PROJECT:         ReactOS kernel
@@ -39,7 +39,7 @@
  *
  * @implemented
  */
-ULONG STDCALL
+SIZE_T STDCALL
 RtlCompareMemory(IN const VOID *Source1,
                  IN const VOID *Source2,
                  IN SIZE_T Length)
